Use user.pk instead of user.username for cache key generation
It is possible to store User objects which have an username which
contains chars that are not ASCII ones. With a cache backend like
Memcached which only allows cache keys constructed from ASCII chars the
cms' cache key generation for user permissions could fail before, this
is prevented by using user.pk.